Update delete und view_all
Update der vorläufig fertigen "delete.php" und "view_all.php"

+++ admin/delete.php
@@ -11,14 +11,21 @@ $sql->execute() or die("fehler");
   }
 
 
-$B_ID = 1;
+if(empty($_GET['b_id'])){
+  die("Es wurde kein Beitrag angegeben der gelöscht werden soll!");
+}
+else{
+  $B_ID = $_GET['b_id'];
+}
+
+
 $sql = $db->prepare("DELETE FROM beitrag WHERE B_ID = ?");
 $sql->execute(array($B_ID)) or die("Löschen Fehlgeschlagen");
 echo "Löschen erfolgreich";
 
 for($i = $B_ID + 1; $i<=$sql_erg; $i++){
     $sql = $db->prepare("UPDATE beitrag SET B_ID = ? WHERE B_ID = ?");
-    $sql->execute(array($i-1, $i)) or die("Fehler UPDATE");
+    $sql->execute(array($i-1, $i)) or die(" Fehler UPDATE");
     
 }
 echo "Datenbank wurde Aktualisiert";
